+++ src/portal/client.ts
@@ -16,8 +16,8 @@
 import { Terminal, ITerminalOptions } from 'xterm'
 import { FitAddon } from 'xterm-addon-fit'
 import { Unicode11Addon } from 'xterm-addon-unicode11'
+import { WebglAddon } from 'xterm-addon-webgl'
 //import { WebLinksAddon } from 'xterm-addon-web-links'
-//import { WebglAddon } from 'xterm-addon-webgl'
 
 //  document elements
 let audioPlay = <HTMLAudioElement>document.getElementById('play')
@@ -267,7 +267,7 @@ function newSession(ev) {
     term.unicode.activeVersion = '11'
     term.open(document.getElementById('terminal'))
     //let's try something new when auto-detection is better suppported
-    //term.loadAddon(new WebglAddon())
+    term.loadAddon(new WebglAddon())
     fit.fit()
     term.blur()
     term.writeln('\x07')